Refactor fetchFromGcs logic
